pyproject.toml
[project]
name = "pyminitel"
version = "0.4.0"
authors = [
    { name="Xenoth", email="xenothvalack@gmail.com" },
]
description = "Minitel Python module"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aenum",
    "pyserial",
    "pillow",
    "redis",
    "nltk"
]

[project.urls]
Homepage = "https://github.com/Xenoth/pyminitel"
Issues = "https://github.com/Xenoth/pyminitel/issues"